clang-tidy Prio 1-3 fixes across 33 files
Prio 1 (correctness):
- cert-flp30-c: replace float loop counters with integer counters in
  GBooleanCollection, GBooleanObject, GDoubleObject
- bugprone-narrowing-conversions: add static_cast in GWorkerT, GRandomFactory
- cert-dcl59-cpp: replace unnamed-namespace constants with inline constexpr
  in GBenchmarkFunctions.hpp
- bugprone-macro-parentheses: GSQUARED wrapped; raiseException E suppressed
  with NOLINT (stream-chained expression cannot be parenthesised)

Prio 2 (safety/robustness):
- concurrency-mt-unsafe: replace exit() with std::terminate() in GLogger;
  NOLINT getenv (called under lock) and exit(0) in --help path in Go2
- bugprone-empty-catch: NOLINT on 13 catch blocks in 8 files that log via
  glogger and terminate via GTERMINATION (opaque to clang-tidy)
- cert-oop54-cpp: self-assignment guards in GPtrVectorT,
  GFixedSizePriorityQueueT, GParserBuilder, GMultiConstraintT (2),
  GRandomDistributionsT, GPlotDesigner (2)
- cppcoreguidelines-pro-type-member-init: default-initialize unset members
  in GFormulaParserT, GPlotDesigner, GNeuralNetworkIndividual,
  GCommonHelperFunctions (struct tm)
- bugprone-assignment-in-if-condition: NOLINT on intentional assignment
  patterns in GWorkerT, GExecutorT, GParameterSet (4×)

Prio 3 (performance):
- performance-noexcept-move: GBaseExecutorT move constructor marked noexcept
- performance-noexcept-swap: swap() declarations and definitions marked
  noexcept in GFormulaParserT, GParameterPropertyParser, GParameterCollectionT

+++ include/courtier/GWorkerT.hpp
@@ -76,7 +76,7 @@ namespace Gem::Courtier
          * supplied by the caller.
          */
         explicit GWorkerT(std::size_t workerId)
-            : m_worker_id(workerId)
+            : m_worker_id(static_cast<std::int32_t>(workerId))
         {
             /* nothing */
         }
@@ -187,7 +187,7 @@ namespace Gem::Courtier
                 {
                     // Retrieve an item and check for its validity. Try again if
                     // we didn't receive a valid item
-                    if (not(p = this->retrieve(m_retrieval_timeout)))
+                    if (not(p = this->retrieve(m_retrieval_timeout))) // NOLINT(bugprone-assignment-in-if-condition)
                     {
                         continue;
                     }
@@ -331,7 +331,7 @@ namespace Gem::Courtier
             {
                 this->process_(p);
             }
-            catch (const g_processing_exception& e)
+            catch (const g_processing_exception& e) // NOLINT(bugprone-empty-catch) — expected; error stored in work item
             {
                 glogger
                     << "In GWorkerT<processable_type>::process():" << std::endl
